Class main working

+++ src/main/kotlin/com/github/eliascoelho911/rnc/converter/RomanNumeralToDecimal.kt
@@ -6,15 +6,15 @@ import com.github.eliascoelho911.rnc.util.RomanNumeralValidator
 class RomanNumeralToDecimal(private val validator: RomanNumeralValidator = RomanNumeralValidator()) {
 
     private val conversionTable =
-        mapOf("I" to 1.0, "V" to 5.0, "X" to 10.0, "L" to 50.0, "C" to 100.0, "D" to 500.0, "M" to 1000.0)
+        mapOf("I" to 1, "V" to 5, "X" to 10, "L" to 50, "C" to 100, "D" to 500, "M" to 1000)
 
     private fun validate(value: String) {
         validator.validate(value)
     }
 
-    fun toConvert(value: String): Double {
+    fun toConvert(value: String): Int {
         validate(value)
-        var total = 0.0
+        var total = 0
         value.forEachIndexed { index, char ->
             val currentDecimalNumber = findDecimalNumber(char.toString())
             val nextDecimalNumber = findNextDecimalNumber(value, index)
@@ -23,12 +23,12 @@ class RomanNumeralToDecimal(private val validator: RomanNumeralValidator = Roman
         return total
     }
 
-    private fun findNextDecimalNumber(value: String, index: Int): Double {
+    private fun findNextDecimalNumber(value: String, index: Int): Int {
         val indexNext = index + 1
-        return if (value.length > indexNext) findDecimalNumber(value[indexNext].toString()) else 0.0
+        return if (value.length > indexNext) findDecimalNumber(value[indexNext].toString()) else 0
     }
 
-    private fun subtractOrSumWithTotal(nextDecimalNumber: Double, currentDecimalNumber: Double, total: Double): Double {
+    private fun subtractOrSumWithTotal(nextDecimalNumber: Int, currentDecimalNumber: Int, total: Int): Int {
         return when {
             currentDecimalNumber < nextDecimalNumber -> total - currentDecimalNumber
             else -> total + currentDecimalNumber
